auth jwt fixed

+++ controllers/profile.js
@@ -3,6 +3,8 @@ const User = require("../models/profile")
 
 const getProfile = async (req, res) => {
   const token = req.headers["authorization"]
+  if (token == "null") return res.json({ message: "Please Login first" })
+
   const id = await getId(token)
   if (id == null) return res.error({ Message: "Not found" })
   const user = await User.findById(id)
@@ -13,6 +15,8 @@ const getProfile = async (req, res) => {
 
 const updateProfile = async (req, res) => {
   const token = req.headers["authorization"]
+  if (token == "null") return res.json({ message: "Please Login first" })
+
   const id = await getId(token)
   const { name, email, password } = req.body
 
@@ -26,6 +30,8 @@ const updateProfile = async (req, res) => {
 
 const deleteProfile = async (req, res) => {
   const token = req.headers["authorization"]
+  if (token == "null") return res.json({ message: "Please Login first" })
+
   const id = await getId(token)
   const user = await User.findByIdAndDelete(id)
